数据库： novel表：添加了作者等级关键字author_rank、小说总字数all_chapter_count novel_chapter表：添加章节字数word_count comment表：添加novel_name，chapter_name这两个关键字 author表：添加follower_count字段 新建：visitor_collect（小说收藏表）、visitor_follow(关注作者表) 前端：修复visitor端，author端的部分排版、数据显示、前后端接口不一致问题 后端：新建的表的相关CRUD以及修改了修改表对应的后端代码；删除了之前的复制工具类，将系统中的更新操作都用动态sql的方式来防止null值被修改进数据库表。

+++ manager-server/src/main/java/com/wang/manager/controller/ManagerController.java
@@ -42,7 +42,7 @@ public class ManagerController {
 
     @DeleteMapping("/delete/{id}")
     @ApiOperation("删除管理员")
-    public Result deleteManager(@PathVariable Long id) {
+    public Result deleteManager(@PathVariable Integer id) {
         log.info("删除管理员请求：ID={}", id);
         return managerService.deleteManager(id);
     }
@@ -73,7 +73,7 @@ public class ManagerController {
     @PostMapping("/updatePassword")
     @ApiOperation("修改管理员密码")
     public Result updatePassword(
-            @ApiParam("管理员ID") @RequestParam Long id,
+            @ApiParam("管理员ID") @RequestParam Integer id,
             @ApiParam("新密码") @RequestParam String newPassword) {
         log.info("修改管理员密码请求：ID={}", id);
         return managerService.updatePassword(id, newPassword);
@@ -88,7 +88,7 @@ public class ManagerController {
 
     @GetMapping("/getNameAndAvatar/{id}")
     @ApiOperation("获取管理员名称和头像")
-    public Result getNameAndAvatar(@PathVariable Long id) {
+    public Result getNameAndAvatar(@PathVariable Integer id) {
         log.info("获取管理员名称和头像请求：ID={}", id);
         return managerService.getNameAndAvatar(id);
     }
