+++ pojo/src/main/java/com/wang/pojo/dto/ManagerDTO.java
@@ -14,7 +14,7 @@ public class ManagerDTO implements Serializable {
     /**
      * 主键ID
      */
-    private Long id;
+    private Integer id;
 
     /**
      * 昵称
@@ -34,7 +34,7 @@ public class ManagerDTO implements Serializable {
     /**
      * 创建者ID
      */
-    private Long createId;
+    private Integer createId;
 
     /**
      * 头像
